requirements.txt
streamlit==1.13.0
Image==1.5.33
numpy==1.22.3
pandas==1.4.2
matplotlib-inline==0.1.2
seaborn==0.11.2
plotly==5.1.0
